fix issue 200 , fix issue 249
git-svn-id: http://nutz.googlecode.com/svn/trunk@1479 423f10f2-e3a4-11dd-a6ea-a32d6b26a33d

+++ src/org/nutz/mvc/upload/UploadInfo.java
@@ -8,8 +8,8 @@ public class UploadInfo implements Serializable {
 
 	public static final String SESSION_NAME = "UPLOAD_INFO";
 
-	public int sum;
+	public long sum;
 
-	public int current;
+	public long current;
 
 }
